+++ changelog.md
@@ -6,7 +6,10 @@
 - Add Docker Ports
 - Add Mesos Discovery
 - Add Docker Volume support
-- Add Environment variables 
+- Add Environment variables
+- Add Restart of Mesos Tasks
+- Add Kill Service 
+- Add Kill single task of a service